Add API key as environment variable; add dotenv gem to read
Sensitive information can now be stored in a .env file in the root directory. The gitignore has been modified to ignore this file.

The dotenv gem has been added to the gemspec as a runtime requirement so that this setup can easily be used in production as well as during development. The new 'require' line in the main module file ensures that environment variables are loaded as early as possible.