Test suite safety: restore card balance after tests
This approach uses a global constant to save the initial balance before running tests, and sets it to 'nil' after using it when the test suite runs.
It could use improvement, by finding another solution besides using a global variable, and by DRYing up the balance checking code. It will also need to be updated when value stores are taken into account, as right now it only considers the principal balance (when checking and re-establishing).