Move API key & base URL to main module
For consistency with Stripe, the API key will be set in code after requiring the Lightrail gem and the API base URL is saved as a variable directly on the main module.

Relevant (basic) specs have been added: to check that the API key is an alphanumeric string, and to check that the base URL points to either the dev or production API. The Connection.connection spec has also been modified to refer to the base URL's new location when checking it on the Faraday connection object.

The spec helper has been modified to set the API key from an environment variable before running anything else in the test suite.