DRY up test suite
Replace repeated fully qualified references to the class being tested in each spec with a named 'subject'. By default RSpec sets the 'subject' to be a new instance of the class the spec is for, but since this project uses class methods it's much more useful to have easy access to the class. Having a named subject also enhances readability.

One spec was handled differently: the spec for Connection currently only tests the .connection method, which returns a connection object (Faraday gem). This spec now uses a 'let' instead of oa 'subject' so as to avoid writing 'connection.connection[...]', but without assuming that the test will only ever test the .connection method ('subject' is meant to be the subject of the whole spec).