Improve Validator methods: raise error if param invalid
Instead of returning false when the param being validated is not valid, raise a LightrailArgumentError. This allows the code in the calling methods to be less verbose (no need for if/else check and raising an error themselves). Error messages include the param in them, for immediate review.

Tests refactored in minor ways:
- Expect erros instead of 'false'
- Break spec into two parts - individual validators (code, transaction_id, etc) and grouped validators that check several individual validators (charge_object etc)
- Use 'let' in each group to reference the LightrailArgumentError class
- L63 create local variable for fund_object so it can easily be included in the failure message