Add Translator to help build HTTP requests from Stripe-like params
Adds a new method that takes in any Stripe-esque param hash (eg charge_object or fund_object) and replaces the terms (keys & values) that Stripe uses with the ones Lightrail uses, eg 'capture' replaced by 'pending' with inverse value.

Accepts a true/false flag to indicate whether the amount/value (if present in the param hash) should be positive or negative when sent to Lightrail. This is a confusing key:value pair, because Stripe's transaction's always expect a positive integer, whereas Lightrail's sometimes expect a positive number and sometimes a negative to distinguish between a funding vs drawdown transaction. Using a flag like this allows the same method to be used for both cases without writing more complicated logic outside of the translation method.