Increase required Ruby to 2.3 to add OpenSSL gem
Stripe requires the use of TLS1.2 or greater. The openssl gem allows for setting the TLS version on the Faraday connection, and it requires at least Ruby 2.3.

Note: set up this way because Ruby relies on the system SSL settings by default; there are likely other ways to handle this than adding a gem.