Refactor hybrid charge creation
Change expected keys in param hash: Stripe source is now :source instead of :stripe_source. This is to facilitate easy adoption of this library, where existing payment flows need to be changed as little as possible. Lightrail payment source can now be any of :code, :lightrail_code, :cardId, :card_id, :lightrail_card_id. Payment will be attempted on the first of those that is found (in that order).

- Add validator for hybrid charge param hash
- Extract verbose/repetitive code into helper functions (eg LR balance checking, determining LR share of transaction, checking for presence of LR & Stripe payment options)
- Add translator for LR charge params from hybrid params
- Add helper methods to translator to modularize methods that need to look at multiple params (updated a couple existing methods to use these)