Change capture/cancel to instance methods
Instead of using class methods on the base LightrailCharge class that require an instance of that class to be passed in as a parameter, make these into instance methods: this way, after calling LightrailCharge.create, the resulting instance can capture!/cancel! itself.

NOTE that these methods actually return new instances of the LightrailCharge class (ie new transactions) and that the original instance remains in the same state with no indication that the transaction has been captured/cancelled. Bangs added to method names to indicate they're doing something unusual (that actually should modify the internal state of the object).