Add skeleton Code & Card classes
These are core concepts in the Lightrail API and need to be part of the wrapper gem. They are specifically needed for balance checking (which doesn't make sense as a 'transaction' or as a 'balance').

This initial version adds balance-checking capabilities for Cards and Codes, plus placeholder charge/fund methods that wrap existing Transaction methods. Happy path specs for all methods are included.

Also: helper method added to Connection class to get the response body of a balance check in a simple way, similar to the helper methods that were implemented to support Transactions.