Add LightrailClient dependency; wrapper classes to sub dir
Dependency on local version of gem added to Gemfile for development; required in main module file.

Removed class definitions for things that are in the LightrailClient gem.

Placeholder 'require' for LightrailClient as a dependency in the Gemspec (pending publication).

Also replace calls to moved Translator methods with calls to new Validator methods (moved when adding classes to LightrailClient gem).