Refactor Translator method: Stripe-style params to LR terms
Instead of having a method that did 90% of the translation and separate methods to handle the last 10% that sometimes varied (depending on whether the transaction amount should be positive or negative), condense into a single method with a true/false flag to handle the positive/negative issue. Also handle cloning the params passed in within the translator method to reduce duplication.