Require Stripe/Lightrail amounts to be passed in to split tender charge
This is an important update to facilitate turnkey implementation.

Instead of automatically calculating the Stripe and Lightrail shares from the total transaction amount and charging as much as possible to Lightrail by default, require the share amounts to be passed in to the default method for creating a split tender charge. This allows for much greater flexibility in checkout, for example allowing the end customer to choose how much of their gift card/account credit balance to apply to the transaction.

A wrapper method has been added that will automatically determine the split, so this functionality is still available (though under a different name).

Basic updates have been made to specs to keep them consistent with these changes.