Split tender method takes Lightrail amount only
Update to conform with interface of other Stripe integration libraries: instead of requiring both the Stripe and Lightrail amounts to be passed into the split tender charge creation method, only accept the Lightrail amount. Since nothing meaningful was being done with the two amounts (there was a check to make sure they added up to the total amount), this just eliminates some unnecessary complexity around calling the method.

There is also still the method that auto-calculates the split based on the maximum that the card can pay.