Refactor split tender validator: check total amount >= lr_share
Covers an edge case where the number passed in for Lightrail's share of the transaction could be greater than the total amount set in the split tender params. Previously, the method would post the charge to Lightrail with the specified Lightrail amount without checking that it made sense wrt the total.

Passing in the Lightrail share to the validator is made optional so that it can still be used in the methods that create/simulate a charge and calculate the split automatically; these methods need to validate the other parameters before they know what the Lightrail amount should be.